+++ docs/homepage/source/conf.py
@@ -13,7 +13,7 @@
 project = 'Boring Math'
 copyright = '2016-2025, Geoffrey R. Scheller'
 author = 'Geoffrey R. Scheller'
-release = '1.2.1' 
+release = '2.0.0' 
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration