+++ src/boring_math/homepage/compatible_versions.py
@@ -1,20 +1,36 @@
 """
-Known Boring Math compatible versions
-=====================================
+Known Boring Math compatible versions with consistent
+Pythonic FP dependencies.
 
-2025-10-TBD - 2.0.0
--------------------
+Coordinated releases
+====================
+
+2025-10-11 - 2.0.0
+------------------
 
 ========================= ===================================== =======
 Name                      PyPI Project                          Version
 ========================= ===================================== =======
 Combinatorics             boring-math-combinatorics             2.0.0  
 Number Theory             boring-math-number-theory             2.0.0  
-Probability Distributions boring-math-probability-distributions 0.9.0  
-Pythagorean Triples       boring-math-pythagorean-triples       0.9.0  
-Recursive Functions       boring-math-recursive-functions       0.9.0  
+Probability Distributions boring-math-probability-distributions 0.8.1  
+Pythagorean Triples       boring-math-pythagorean-triples       0.8.3  
+Recursive Functions       boring-math-recursive-functions       1.0.0  
 Special Functions         boring-math-special-functions         0.1.0  
 ========================= ===================================== =======
+
+The Pythonic FP external dependencies these brought in on 2025-10-11 were
+
+.. code:: console
+
+    $ pip list | grep pythonic
+    pythonic-fp-booleans                  2.0.0
+    pythonic-fp-circulararray             6.0.0
+    pythonic-fp-fptools                   5.1.2
+    pythonic-fp-gadgets                   3.1.0
+    pythonic-fp-iterables                 5.1.2
+
+For which all current applicable Boring Math tests passed.
 
 ----
 
@@ -29,5 +45,18 @@ Probability Distributions boring-math-probability-distributions 0.8.0
 Pythagorean Triples       boring-math-pythagorean-triples       0.8.2  
 Recursive Functions       boring-math-recursive-functions       0.8.1  
 ========================= ===================================== =======
+
+The Pythonic FP external dependencies these brought in on 2025-11-10 were
+
+.. code:: console
+
+    $ pip list | grep pythonic
+    pythonic-fp-booleans                  2.0.0
+    pythonic-fp-circulararray             6.0.0
+    pythonic-fp-fptools                   5.1.2
+    pythonic-fp-gadgets                   3.1.0
+    pythonic-fp-iterables                 5.1.2
+
+For which all current applicable Boring Math tests passed.
 
 """