+++ src/boring_math/homepage/overview.py
@@ -2,10 +2,10 @@
 The overall effort's name is
 `Boring Math <https://pypi.org/project/boring-math/>`_.
 
-- it is a collection of my mathematical hobby projects
-- each project's PyPI and GitHub names begin with *"boring-math-"*
-- each project implements a top-level package in the ``boring_math`` Python namespace
-- for links to all the Boring Math project PyPI and GitHub repos `click here
-  <https://github.com/grscheller/boring-math/blob/main/README.md>`_
+- It is a collection of my mathematical hobby projects.
+- Each project's PyPI and GitHub name begins with *"boring-math-"*.
+- Each project implements a top-level package in the ``boring_math`` Python namespace.
+- Here are links to all the Boring Math project `PyPI and GitHub repos
+  <https://github.com/grscheller/boring-math/blob/main/README.md>`_.
 
 """